requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
audioop-lts==0.2.2
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cramjam==2.11.0
faiss-cpu==1.12.0
fastapi==0.116.1
fastparquet==2024.11.0
ffmpy==0.6.1
filelock==3.19.1
fsspec==2025.7.0
ftfy==6.3.1
gradio==5.43.1
gradio_client==1.12.1
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kagglehub==0.3.12
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
open_clip_torch==3.1.0
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
ruff==0.12.10
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
